fixed setup

+++ matcher/output.py
@@ -202,7 +202,8 @@ def update_csv_betfair(
 def reset_csv():
     if not os.path.isdir("stats"):
         os.mkdir("stats")
-    shutil.copyfile(".env.template", ".env")
+    if not os.path.isfile(".env"):
+        shutil.copyfile(".env.template", ".env")
     now = datetime.now().strftime("%d-%m-%Y")
     RETURNS_HEADER = "date_of_race,horse_name,bookie_odds,venue,ew_stake,balance,rating,expected_value,expected_return,win_stake,place_stake,win_odds,place_odds,betfair_balance,max_profit,is_lay,win_matched,lay_matched,arbritrage_profit,place_payout,balance_in_betfair,current_time"
     RETURNS_BAK = os.path.join(BASEDIR, "stats/returns-%s.csv" % now)
@@ -210,11 +211,15 @@ def reset_csv():
 
     if os.path.isfile(RETURNS_CSV):
         create_new_returns = input(
-            "Create new return.csv? (Recommended for new user) [Y]/n "
+            "Create new return.csv? (Recommended for new user) Y/[n] "
         ).lower()
 
-    if create_new_returns != "n":
+    if create_new_returns == "y":
         if os.path.isfile(RETURNS_CSV):
+            count = 2
+            while os.path.isfile(RETURNS_BAK):
+                RETURNS_BAK = f"{RETURNS_BAK}({str(count)})"
+                count += 1
             os.rename(RETURNS_CSV, RETURNS_BAK)
         with open(RETURNS_CSV, "w", newline="") as returns_csv:
             returns_csv.write(RETURNS_HEADER)
